refactor(fbuild-packages): split library/esp32_framework.rs into submodules to satisfy LOC gate
